apps/ff14-submarines/requirements.txt
mss==9.0.1
pillow==10.4.0
pytesseract==0.3.13
google-api-python-client==2.138.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
